+++ src/components/chat/chat.component.tsx
@@ -1,4 +1,4 @@
-import { Col, Row } from "react-bootstrap";
+import { Col } from "react-bootstrap";
 import { useChatbot } from "../../hooks/use-chatbot.hook";
 import { useFocusNewMessage } from "../../hooks/use-focus-new-message.hook";
 import { useMessages } from "../../hooks/use-messages.hook";
